fix: don't refresh monitor every state change, lower speaking epoch

+++ crates/orbolay-core/src/user.rs
@@ -44,7 +44,7 @@ pub struct User {
   pub streaming: bool,
   pub camera: bool,
   pub volume: f32,
-  pub speaking_epoch: u64,
+  pub speaking_epoch: f32,
 }
 
 use crate::payloads::WsVoiceState;
@@ -62,7 +62,7 @@ impl From<WsVoiceState> for User {
       streaming: val.streaming.unwrap_or_default(),
       camera: val.camera.unwrap_or_default(),
       volume: 100.,
-      speaking_epoch: 0,
+      speaking_epoch: 0.,
     }
   }
 }
@@ -104,7 +104,7 @@ impl From<RpcVoiceState> for User {
       streaming: false,
       camera: false,
       volume: val.volume,
-      speaking_epoch: 0,
+      speaking_epoch: 0.,
     }
   }
 }
